devtools/shared/webconsole/test/test_object_actor_native_getters_lenient_this.html

<!DOCTYPE HTML>
<html lang="en">
<head>
  <meta charset="utf8">
  <title>Test that WebIDL attributes with the LenientThis extended attribute
         do not appear in the wrong objects</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="common.js"></script>
  <!-- Any copyright is dedicated to the Public Domain.
     - http://creativecommons.org/publicdomain/zero/1.0/ -->
</head>
<body>
<p>Test for the native getters in object actors</p>

<script class="testbody" type="text/javascript">
SimpleTest.waitForExplicitFinish();

function startTest()
{
  removeEventListener("load", startTest);

  attachConsoleToTab(["ConsoleAPI"], onAttach);
}

function onAttach(aState, aResponse)
{
  onConsoleCall = onConsoleCall.bind(null, aState);
  aState.client.on("consoleAPICall", onConsoleCall);

  let docAsProto = Object.create(document);

  top.console.log("hello", docAsProto);
}

function onConsoleCall(aState, aPacket)
{
  info("checking the console API call packet");

  checkConsoleAPICall(aPacket.message, {
    level: "log",
    filename: /test_object_actor/,
    functionName: "onAttach",
    arguments: ["hello", {
      type: "object",
      actor: /[a-z]/,
    }],
  });

  aState.client.off("consoleAPICall", onConsoleCall);

  info("inspecting object properties");
  let args = aPacket.message.arguments;
  onProperties = onProperties.bind(null, aState);

  let client = new ObjectClient(aState.dbgClient, args[1]);
  client.getPrototypeAndProperties(onProperties);
}

function onProperties(aState, aResponse)
{
  let props = aResponse.ownProperties;
  let keys = Object.keys(props);
  info(keys.length + " ownProperties: " + keys);

  is(keys.length, 0, "number of properties");
  keys = Object.keys(aResponse.safeGetterValues);
  is(keys.length, 0, "number of safe getters");

  closeDebugger(aState, function() {
    SimpleTest.finish();
  });
}

addEventListener("load", startTest);
</script>
</body>
</html>
